Feature/deb dist problems

+++ ci/scripts/linux-cross/after_success.sh
@@ -18,4 +18,8 @@
 # along with Flom.  If not, see <http://www.gnu.org/licenses/>.
 #
 
-docker run --rm -v $(pwd):/source -v $(pwd)/build:/build builder -c "bash /package.sh"
+mkdir -p build/dist
+
+if ${BUILD_PACKAGE:-true}; then
+  docker run --rm -v $(pwd):/source -v $(pwd)/build:/build builder -c "bash /package.sh"
+fi
